+++ extensions/uptime-robot/src/monitors.tsx
@@ -9,13 +9,26 @@ import {
   MONITOR_INTERVALS,
   MONITOR_TYPES,
 } from "./config";
-import { Action, ActionPanel, Alert, confirmAlert, Form, Icon, Keyboard, List, showToast, Toast, useNavigation } from "@raycast/api";
+import {
+  Action,
+  ActionPanel,
+  Alert,
+  Color,
+  confirmAlert,
+  Form,
+  Icon,
+  Keyboard,
+  List,
+  showToast,
+  Toast,
+  useNavigation,
+} from "@raycast/api";
 import { ErrorResponse, Monitor, NewMonitor } from "./types";
 import { useEffect, useState } from "react";
 import useUptimeRobot from "./lib/hooks/use-uptime-robot";
 import unixToDate from "./lib/utils/unix-to-date";
 import { hasDayPassed } from "./lib/utils/has-day-passed";
-import deleteMonitor from "./lib/utils/delete-monitor";
+import { deleteMonitor } from "./lib/api";
 
 type Pagination = {
   offset: number;
@@ -39,7 +52,7 @@ export default function Monitors() {
     isLoading,
     pagination,
     data: monitors,
-    mutate
+    mutate,
   } = useFetch(
     (options) =>
       API_URL +
@@ -84,27 +97,25 @@ export default function Monitors() {
 
   async function confirmAndDeleteMonitor(monitor: Monitor) {
     const options: Alert.Options = {
-      icon: Icon.Trash,
+      icon: { source: Icon.Trash, tintColor: Color.Red },
       title: "Are you sure?",
       message: `Do you really want to delete ${monitor.friendly_name}? This can't be undone.`,
       primaryAction: {
         title: "Delete",
-        style: Alert.ActionStyle.Destructive
-      }
-    }
+        style: Alert.ActionStyle.Destructive,
+      },
+    };
     if (await confirmAlert(options)) {
       const toast = await showToast(Toast.Style.Animated, "Deleting monitor", monitor.friendly_name);
       try {
-        await mutate(
-          deleteMonitor(monitor.id), {
-            optimisticUpdate(data) {
-              return data.filter(m => m.id !== monitor.id);
-            },
-            shouldRevalidateAfter: false
-          }
-        )
+        await mutate(deleteMonitor(monitor.id), {
+          optimisticUpdate(data) {
+            return data.filter((m) => m.id !== monitor.id);
+          },
+          shouldRevalidateAfter: false,
+        });
         await setMonitors({
-          monitors: monitors.filter(m => m.id!==monitor.id),
+          monitors: monitors.filter((m) => m.id !== monitor.id),
           updated_at: new Date(),
         });
         toast.style = Toast.Style.Success;
@@ -125,8 +136,11 @@ export default function Monitors() {
           title={monitor.friendly_name}
           icon={MONITOR_ICONS[monitor.status]}
           accessories={[
-            { icon: Icon.Redo },
-            { text: `${monitor.interval / 60} min` },
+            {
+              icon: Icon.Redo,
+              text: `${MONITOR_INTERVALS[monitor.interval]}`,
+              tooltip: `Checked every ${MONITOR_INTERVALS[monitor.interval]}`,
+            },
             { date: unixToDate(monitor.create_datetime) },
           ]}
           actions={
@@ -137,7 +151,19 @@ export default function Monitors() {
                 title="Add New Monitor"
                 target={<AddNewMonitor onMonitorAdded={() => setExecute((prev) => !prev)} />}
               />
-              <Action icon={Icon.Trash} title="Delete Monitor" onAction={() => confirmAndDeleteMonitor(monitor)} style={Action.Style.Destructive} shortcut={Keyboard.Shortcut.Common.Remove} />
+              <Action.OpenInBrowser
+                icon="up.png"
+                title="Open in Dashboard"
+                url={`https://dashboard.uptimerobot.com/monitors/${monitor.id}`}
+                shortcut={Keyboard.Shortcut.Common.Open}
+              />
+              <Action
+                icon={Icon.Trash}
+                title="Delete Monitor"
+                onAction={() => confirmAndDeleteMonitor(monitor)}
+                style={Action.Style.Destructive}
+                shortcut={Keyboard.Shortcut.Common.Remove}
+              />
             </ActionPanel>
           }
         />